fix(hooks): refuse declarations made after a bare `--`, and state what binds
Two items from adversarial review of #2495.

**1. End-of-options position blindness — a LIVE bypass on the acli path.**

`flag_values()` did not honour `--`, so the guard credited a `--label` that
the CLI would never apply:

    gh issue create --title x -- --label status:ready

The review flagged this as unexploitable on gh, and on gh that is right: gh
2.96.0 rejects it outright ("unknown arguments [--label status:ready]") before
any API call. But the review's real point was that gh's strictness cannot be
assumed to generalize, and running the other installed CLI proved it:

    acli jira workitem create --summary probe --project ZZZNOTREALPROJ \
      --type Task -- --status Ready --label status:ready
    -> Authenticated site: geminisportsanalytics.atlassian.net
    -> Error: No project could be found with key 'ZZZNOTREALPROJ'.

acli parses straight past `--` and proceeds to a server-side project lookup.
With a real project key it would have CREATED the work item, with the trailing
`--status` silently unapplied and the guard's declaration check satisfied by a
token that never reached the item. That is a live bypass on the JIRA path, not
a theoretical one.

`jira`, `linear`, `http`, and `wget` are not installed here, so their
end-of-options behavior is UNVERIFIED. The guard now fails closed on every
path rather than inheriting gh's strictness — a vendor-neutral control cannot
assume the strictest vendor is representative, and this is the case that
proves it. The tests and the rule say which CLIs were actually run and which
were not, instead of implying uniform coverage.

The fix is the same invariant as the previous commit's, one position over:
check the artifact. A post-`--` token is an operand, not a flag, so it cannot
reach the created item — and so it cannot declare anything, including the
`[lisa-human-gate]` marker. Ported to the OpenCode TS guard and verified
against the same cases. 5 new regressions, 4 of which fail against HEAD~.

**2. Naming which half of this work actually binds.**

The two defects in this branch do NOT bind equally, and this session's whole
thesis is that prose and executable controls do not bind equally, so saying
otherwise here would be self-refuting.

`ready-role-filing` is a genuine executable control: a hook that exits 2 and
physically refuses the tool call.

`runtime_behavior_change` is not. Nothing refuses anything; enforcement is a
validator agent reading SKILL.md and choosing to apply it — the same rung that
failed 13 filings out of 13. The regression suite pins the WORDING across all
six generated roots, so the contract cannot be silently deleted, softened, or
lost in a regeneration. It cannot catch a validator that documents the rule
and ignores it at runtime.

That is the ceiling rather than a shortcut: SKILL.md IS the validators'
execution substrate, so there is no compiled artifact underneath to hook. The
honest claim is that absence-as-`false` was undecidable and is now DECIDABLE —
real, and strictly weaker than ENFORCED. The rule now says so, and tells
readers to treat a verdict resting on it as evidence rather than proof until a
deterministic checker exists.

Work-Item: CodySwannGT/lisa#2486


Work-Item: CodySwannGT/lisa#2486
